Fix - Artifactory deploy proprieties are not saved on deployed artifacts
By adding back the original usage of gatherBuildInfoParams
all properties that start with "artifactory.deploy" are being set
on the deployed artifacts.
